travis warnings

+++ src/tools/csg_resample.cc
@@ -167,7 +167,7 @@ int main(int argc, char** argv)
         if (!vm.count("nocut")) {
             // determine vector size
             int minindex=-1, maxindex=-1;
-            for (size_t i=0; i<in.x().size(); i++) {
+            for (int i=0; i<in.x().size(); i++) {
                 if(in.x(i)<sp_min) {
                     minindex = i;
                 }
@@ -232,10 +232,10 @@ int main(int argc, char** argv)
     der.GenerateGridSpacing(min, max, step);
     der.flags() = std::vector<char>(der.flags().size(), 'o');
 
-    unsigned int i=0;
+    int i=0;
     for(i=0; out.x(i) < in.x(0) && i<out.size(); ++i);
 
-    unsigned int j=0;
+    int j=0;
     for(;i < out.size(); ++i) {
         for(; j < in.size(); ++j)
             if(in.x(j) >= out.x(i)  || fabs(in.x(j)-out.x(i) ) < 1e-12) // fix for precison errors
